backend/trends/requirements.txt
Flask
sqlalchemy
psycopg2-binary
alembic>=1.2
apscheduler
pytrends
Flask-Caching
PyYaml
coloredlogs